+++ client/src/redux/action_types.js
@@ -12,3 +12,10 @@ export const REGISTER_SUCCESS = 'REGISTER_SUCCESS'
 export const REGISTER_FAIL = 'REGISTER_FAIL'
 
 export const LOGOUT = 'LOGOUT'
+
+//======================================
+// ALERT
+//======================================
+
+export const SET_ALERT = 'SET_ALERT'
+export const REMOVE_ALERT = 'REMOVE_ALERT'